+++ packages/vc-delegation-engine/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @docknetwork/vc-delegation-engine
 
+## 1.0.3
+
+### Patch Changes
+
+- Early out failure on delegation verify if missing credentials from chain
+
 ## 1.0.2
 
 ### Patch Changes